[windows/install] - refactor + fixed windows install commands
