+++ preinstall-scripts/windows.commands.js
@@ -1,0 +1,22 @@
+const refreshPath =
+  "[System.Environment]::SetEnvironmentVariable('Path', [System.Environment]::GetEnvironmentVariable('Path','Machine') + ';' + [System.Environment]::GetEnvironmentVariable('Path','User')); refreshenv";
+
+const commands = [
+  'Set-ExecutionPolicy Bypass -Scope Process -Force',
+  '[System.Net.ServicePointManager]::SecurityProtocol = [System.Net.ServicePointManager]::SecurityProtocol -bor 3072',
+  "iex ((New-Object System.Net.WebClient).DownloadString('https://chocolatey.org/install.ps1'))",
+  'choco install imagemagick.app -PackageParameters LegacySupport=true -PackageParameters installDevelopmentTools=true -y',
+  'choco install nodejs-lts python ghostscript qpdf pandoc mupdf -y',
+  'choco install tesseract --pre -y',
+  'curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py',
+  refreshPath,
+  'python get-pip.py',
+  'pip install pdfminer.six camelot-py[cv]',
+  refreshPath,
+];
+
+module.exports = [
+  'powershell.exe',
+  '-Command',
+  `Start-Process powershell -Wait -Verb RunAs "${commands.join('; ')}"`,
+];
